Avoid calls from SYSTEM_ADDRESS with signed messages
Calls from SYSTEM_ADDRESS are not stored on the blockchain, so the same
calls must be made by every node.  This breaks the current contracts,
which rely on different nodes to make different calls, and use digital
signatures to distinguish the callers.

Instead, just use ordinary calls, and get the caller from `msg.sender`.
This does consume gas, but that will be fixed later.